GH-93: Rename Article List Plugin (add Static)
- Add "(Static)" to "Article List" plugin human name.
- Add `static_` to `article_list` plugin code name.
- Move "(Static)" in "Article Preview" plugin human name.
