+++ taccsite_cms/settings.py
@@ -243,7 +243,7 @@ INSTALLED_APPS = [
     'taccsite_cms.contrib.taccsite_sample',
     'taccsite_cms.contrib.taccsite_blockquote',
     'taccsite_cms.contrib.taccsite_offset',
-    'taccsite_cms.contrib.taccsite_article_list',
+    'taccsite_cms.contrib.taccsite_static_article_list',
     'taccsite_cms.contrib.taccsite_static_article_preview',
 ]
